avoid using Scalar::Util functions in List::Util tests
This ensures that changes to one won't impact the test for the other.
It also prepares for a potential future where the modules are split into
separate dists.  This also makes it easier to port tests to
List::Util::PP, which wants to avoid all direct dependencies on
Scalar::Util.